+++ src/love/Types/Audio/SoundData.h
@@ -2,7 +2,6 @@
 #define SRC_LOVE_TYPES_AUDIO_SOUNDDATA_H_
 
 #include <string>
-#include "physfs.h"
 #include "audio/audio_mixer.h"
 
 namespace love {
@@ -34,6 +33,8 @@ class SoundData {
 	 * Gets the current volume of the Source.
 	 *
 	 * @return The volume of the source, from 0.0f to 1.0f.
+	 *
+	 * @see setVolume
 	 */
 	float getVolume();
 
@@ -43,6 +44,8 @@ class SoundData {
 	 * @param volume The volume to set the source at, from 0.0f to 1.0f.
 	 *
 	 * @return The SoundData object iself, to allow for method chaining.
+	 *
+	 * @see getVolume
 	 */
 	SoundData& setVolume(float volume);
 
@@ -55,24 +58,29 @@ class SoundData {
 
 	/**
 	 * Returns whether the Source will loop.
+	 *
+	 * @see setLooping
 	 */
 	bool isLooping();
 
 	/**
 	 * Set whether the Source should loop.
+	 *
+	 * @see isLooping
 	 */
 	SoundData& setLooping(bool loop);
+
+	// The audio callback for when a sound finishes.
+	static void audioCallback(audio_mixer_sound_t* sound, unsigned reason);
 
 	// Properties
 	bool m_playing = false;
 	bool m_loop = false;
 	float m_volume = 1.0f;
+	void* buffer = NULL;
 	audio_mixer_sound* m_sound = NULL;
-	// TODO(RobLoach): Make voice a vector.
+	// TODO(RobLoach): Make voice a vector?
 	audio_mixer_voice_t* m_voice = NULL;
-
-	// The audio callback for when a sound finishes.
-	static void audioCallback(audio_mixer_sound_t* sound, unsigned reason);
 };
 
 }  // namespace Audio