+++ src/love/audio.h
@@ -10,6 +10,8 @@ namespace love {
 
 /**
  * Provides an interface to output sound to the user's speakers.
+ *
+ * @see love.sound
  */
 class audio {
 	public:
@@ -29,6 +31,8 @@ class audio {
 	 * @param type ("static") The type of audio source to load ("static", "stream"). Optional.
 	 *
 	 * @return The new SoundData if it loaded correctly, NULL otherwise.
+	 *
+	 * @see love.sound.newSoundData
 	 */
 	SoundData* newSource(const std::string& filename, const std::string& type);
 	SoundData* newSource(const std::string& filename);