+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to [ChaiLove](https://github.com/RobLoach/ChaiLove) will be 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.25.0 - 2018-09-08
+### Features
+- Added Cheat support
+  - `cheatset(index, enabled, code)`
+  - `cheatreset()`
+- Added `.ogg` audio support
+
+### Fixes
+- Fixed save state loading
+- Fixed `love.filesystem.mount()` logic
+
 ## 0.24.0 - 2018-09-03
 ### Features
 - Update to use libretro-common audio